applets/digital-clock: Fix sizing and layouts of CalendarView full representation
It makes sizes and paddings more consistent, and fixes RTL issues such
as manually mirroring left/right padding properties.

Top level layout is now simpler thanks to the anchors. But width still
has to be manually set to round(parent.width/2) to keep it pixel-perfect because anchors don't care and binding to e.g. horizontalCenter would
result in fractional positioning without any easy way to snap it back.

Top-level custom property "paddings" was increased from 1 smallSpacing
to 2 of them, since that's what really consistent with the calendar view
on the other side, and it looks better for border-touching items anyway.