shell: Port DesktopView to layer-shell
This ports the DesktopView from the plasma-shell protocol to the
layer-shell protocol.

There's one small caveat about the layer-shell protocol: the layer
surface is bound to the output, in order to move a DesktopView between
outputs, the DesktopView's visibility must be toggled to re-create the
layer surface with the new output.