users kcm: hack around indexing confusion in swipeview
swipeview gets confused over which index it is on causing it to
automatically switch to an incorrect index. I know it has to do with
Container internals but haven't quite managed to track it down in
reasonable time. there are various Qt bug reports to do with
currentIndex handling being buggy though.

intermediate index resets seem to make the view behave accordingly so
let's apply one when setting the current index.

BUG: 439081