users kcm: set an implicitwidth on PicturesSheet
there appears to be a hidden binding loop between the width and height
(or so the engine says), presumably because of how nothing here has a
concrete size. so, let's help things along by breaking the loop and
making our preferred width the implicitwidth. this has much the same end
result but actually works consistently.