applets/notification: fix history view vertical lines
Right now we use a thickened version of the Plasma theme's vertical line
SVG for the lines in grouped history items. However this is an abuse of
the SVGs, which were not necessarily designed to be thickened in that
way. The result looks fine for some themes where the SVG really is just
a plain vertical line, but does not look good for themes where the
graphic is fancier--for example, Oxygen.

Since the look we want here is just a thick line (AKA a rectangle),
let's just create one ourselves to avoid drama with the Plasma theme.

BUG: 416386
FIXED-IN: 5.27.5


(cherry picked from commit 591ce1f1f7e9f99d3854975e2a6d605ac4ecb44c)