applets/notifications: Set text color to ensure visibility with all color themes
Text in the body of notifications is displayed in a TextArea without a background.
This causes an issue in themes that expect text in textareas to be a different
color than ordinary labels; the notification text can end up unreadable.

Manually setting the text color fixes the issue.


(cherry picked from commit 68f8dbddc9bf79396a6200200042541f62043647)