Fix Klipper empty prevention on wayland
When an application terminates, we receive a signal for the Selection and restore it. While restoring it,
we also receive a signal for the Clipboard, but don't handle it since we're currently locked due to the Selection.

This patch fixes that by creating separate locks for Selection and Clipboard.

BUG: 468422
(cherry picked from commit c6d52910ebc1fb1abd89d286aec9c24ca102db77)