kcms/notifications: Move per-event config into main page
Configure the notification events inline in the notifications
settings of every application, instead of an external dialog
provided by KNotifyConfig.

This adds a direct dependency to libcanberra, to play the
notification sound previews, but it was already indirectly
dependent via KNotifyConfig. It also drops Phonon support
for sound previews

To correctly implement the expected KCM behaviors for events
(saving, restoring and defaults), the EventData struct is
replaced with a more handy KConfigXT EventSettings object.