krunner: Show suggestions as grayed out text, press arrow-right to apply
This way, we do not apply a suggestion without asking the user for it.
Using the grayed out text, it is made clear to the user what the
suggestion is.

Because the old behavior is desirable for existing users, it is still
kept as an option. The code required for it is minimal anyway.

FEATURE: 413709
FIXED-IN: 6.0