kcms/region_language: use alternating background for list items
These list items have content on both sides, separated by a potentially
very large area of whitespace. This can make it hard to distinguish
at-a-glance which items on the right side line up with which items on
the left side.

The typical way to improve usability for such a thing is to use
alternating background colors to separate list items in a way that
isn't super visually heavy. Let's do it here.

BUG: 469102
FIXED-IN: 6.0