Revert "Re-add Force Font DPI spinbox on Wayland"
In the past we re-added this because people were still legitimately
using it to work around blurry scaling. But that's now fixed in Qt 6 for
Qt apps, and there's a path forward for other ones too. So let's remove
it again so reduce the number of methods that exist to scale things,
which we know causes confusion.

This reverts commit 62842ac6fd5571394e38360b345a7ad3a1187273.

Closes https://invent.kde.org/plasma/plasma-desktop/-/issues/62.