Klipper config: set UI state when loading the same way as when editing
When "Keep the selection and clipboard the same" is unchecked,
some widgets are disabled. But that was only applied when
unchecking the checkbox after opening the dialog, not when the
checkbox was initially (dialog loading time) unchecked due to
settings. Now, widgets that should be disabled at dialog loading
time *are* disabled.


(cherry picked from commit 1e25653e368555fc61dff4c6d604df1dcec7b55c)