Address some QStringLiteral data duplication
Each use produces a new UTF-16 array that contains the string data.
This duplicates the data as many times as you “call” QStringLiteral
with the same argument. It does so even within a single translation
unit. Common C string literals, on the other hand, are allowed to
share a single memory location.

Ref: https://www.kdab.com/qstringview-diaries-advances-qstringliteral/