wallpapers/image: resize svg wallpaper to requested size
QImage is rasterized, which means it needs a defined size. If we need to
resize a svg image, the svg image should be loaded by QSvgRenderer and then
QPainter class is used to draw the image onto a new QImage with the desired size.

Plasma::Svg can help wrap the whole process and provide a cached image
if the wallpaper is loaded again.

BUG: 469294
FIXED-IN: 5.27.6

(cherry picked from commit a3d0491deebb460db5a033c4a9679e6a5452372e)