Add a KDED module to ask for donations once a year
KDE's recent experience with donations has been that when we ask, people
donate. Our 2022 fundraiser brought in 35'000€, and the 2023-2024 one
brought in over 140'000€. And these were advertised only through KDE's
media channels, not in the software itself.

Income like this is important for KDE e.V.'s financial sustainability
and ability to grow and expand initiatives, but it still pales in
comparison to current and expected future expenses.

Other popular software-based nonprofits such as Thunderbird and
Wikimedia ask for donations yearly in the UI of their software, and see
huge returns from it. Accordingly, this commit implements a yearly
system notification in Plasma that asks the user to make a donation.
It's shown once in December, and then won't appear again for a year.

It's implemented as an on-by-default KDED module, meaning distros and
users can permanently disable it using the normal means of disabling
KDED modules (i.e. kcm_kded).