applets/batterymonitor: Fix wrong tooltip when plugged in but discharging
In this case, the tooltip subtext line would still say "X time until
fully charged" because it was not checking to make sure that the battery
was in fact actually charging. This commit makes it do so, so now it
will fall through to the next condition that displays text suitable for
a battery that's discharging.

BUG: 470632
FIXED-IN: 5.27.6