applets/notifications: ensure item exists before querying its properties
In NotificationItem.qml, we query a loader's item's implicit height only
when the loader is active. But it's possible for the loader to be active
and not yet have an item, since the item may take a moment to load. So
let's guard that by also making sure the item is set before we query its
properties.