lookandfeel/osd: Fix spacing in certain scenarios
This commit fixes spacing when only a heading is visible (it has only
right margin) and when an icon isn't shown when showing progress (the
progress bar is lacking left margin).

There's no more elegant way I can figure to achieve this. Using empty
items with a set width causes the layout's spacing to also apply to
those items.