remove the force qt scaling env var
Now we use Qt scaling everywhere, to reduce the number of different code
paths and visual differences between the X11 and Wayland sessions. In the
past we didn't do this because there were too many bugs, but almost all
of them have been fixed now.

The one thing this is known to regress is the mixed-DPI multimonitor use
case. But that use case is already not supported on X11, so our story
remains the same: if you have a mixed-DPI multimonitor setup, use Wayland.

BUG: 356446
FIXED-IN: 6.0