krunner: only replace with completion text on right arrow when relevant
We should only do this when the cursor position is at the end of the
typed text and there is faded completion text after it. At all other
times, the right arrow key should do normal navigation among characters,
or else pressing it can produce weird and unexpected results.

BUG: 472007
FIXED-IN: 6.0