applets/digital-clock: use ScrollViewKCM for timezone page
This ensures that only the inner scrollable view will scroll, rather
than the rest of the page.

In the future we can make this scrollview frameless like our KCMs, but
first we'll need to add a way to remove the built-in margins for all
page content that get added by this dialog window.