applets/digital-clock: don't remove highlight effect on timezone page
It may not look the best, but removing it creates an accessibility
problem because you can't see what has focus anymore. It also amounts to
a minor usability issue since it's no longer clear that the whole list
item is selectable. The appearance can and should be fixed at a lower
level so we aren't tempted to get rid of it.

But until that happens, we do still need to work around the other
usability issue that the radio button disappears for highlighted items.
So as a compromise, use the hover effect instead.