applets/notifications: add workaround for QTBUG-100392
Citing the bugreport:

```
Qt/QML crashes at runtime when an animation is started for the second time after caching its targets' properties, but one or more of those targets are actually destroyed by the time it restarts.

It is likely because the list<> property type does not emit any on*Changed() signal when any of its items are destroyed — either manually by invoking QtObject::destroy() method, by an Instantiator when it removes its managed objects, or in any other way. Lists start showing null values after control returns to the even loop (so that Qt finishes actually destroying an object and cleans up all references it can find), but according to the backtrace Animation does some property caching, and so it tries to access freed memory leading to Segmentation fault.
```

Ref: https://bugreports.qt.io/browse/QTBUG-100392
Ref: https://invent.kde.org/qt/qt/qtdeclarative/-/commit/8a74155411d7a22bb8efa405fe59a681e47e6b45

BUG: 468180
FIXED-IN: 5.27.7