runners/services: Optimize by excluding NoDisplay apps early
By filtering them out in the prepare slot, we do not need to check
for the noDisplay property for every character typed and in multiple
places.
We also allocate less memory, because irrelevant services are not kept
in the internal list for the match session.

The optimizations from https://invent.kde.org/frameworks/kservice/-/merge_requests/154
benefit this also