krunner: Add indicator/toggle for singleRunnerMode
This mode was recently exposed as a command-line option to KRunner.
The selected runner's name is part of the placeholder text, but when
typing a query you can no longer see it. Also, the only way to quit this
mode is to reopen KRunner.

Instead, there is a simple button with the runner's icon indicating that
we only query this one runner. When it is pressed, we switch back to
normal mode and hide the button.