runners/calculator: Do not call QObject that lives in another thread
Run is called in the main thread, meaning we should only access the
match and context. Meaning we should not get the text copied in the
engine, but rather read the match text and copy it directly to the
clipboard.

The only difference might a change in the delimiters, but now we copy
the text that the user sees. IMHO that is the best approach anyway.

BUG: 472469