shell: move `forceActiveFocus` to `PanelView`
This allows to focus on the first child widget in a panel by setting
containment status directly.

This adds `savePreviousWindow()` to save the previously focused window,
to be called in `PanelView::refreshStatus`. When the previous window is
not reset, calling savePreviousWindow() will not do anything, so when
there are more than one panel, it will not incorrectly save the wrong
window (panel).

CCBUG: 455398
CCBUG: 352476
CCBUG: 453166