Add KDED module for adjusting the system time zone based on location
We don't need accurate location data, so it just talks to the KDE geoip endpoint
directly and adjusts the timezone accordingly.

It does that only on a network with full connectivity (ie. not behind a portal),
that is not metered and not connected to a VPN (to avoid using the VPN gateway's
timezone inadvertendly).

It only does it max every 25 hours (deliberately +1 to avoid checking at the
same time of the day), unless connected to a new connection to ensure prompt update
when arriving at a new location. A "refresh()" DBus method is available to let
the datetime KCM trigger a manual check.

Checks are still delayed by a random bit to spread out traffic.

The module is off-by-default to not do unsolicited network requests and can
be enabled from datetime settings.