Avoid losing clipboard data if we exit abnormally
If we have just changed the settings from non-persistent to persistent
clipboard history, we should save immediately, in case we are unable to
save normally at exit.

Old behaviour: when settings are changed from Save to Don't Save, wipe
the history from disk (privacy) but when they are changed back, don't
bother saving the data to disk - normally this is fine, since we save
when plasmashell exits, but if it were to crash before being saved the
data would be lost.

New behaviour: save or wipe the clipboard history (as applicable)
whenever the setting is changed, not only when it is changed from Save
to Don't Save.