OSD: Fix size calculation for progress value
Use a common function for formatting the percent, so placeholder 100% label will
use the exact formatting used by the label next to the progress bar.

Also use font advanceWidth, which is the appropriate means of calculating the
bounding box of text in this case. Also ceil it to avoid subpixel alignment.

And to be absolutely sure disable word wrapping.

BUG: 469576