kcms/soundtheme: Better compact mode calculation
Slightly adjust the width calculation to position the demo
play button on top and avoid edge cases (ex Noto Sans 11pt).

This mode is however a small workaround as we cannot set a
minimum size for the cards. Ideally, it can be dropped if
we adopt non-resizable cards for the KCMs.

BUG: 472899
FIXED-IN: 6