kcms/soundtheme: Provide better info for the "freedesktop" theme
The "freedesktop" theme is the fallback sound theme according to
the specification. It is equivalent to "hicolor" for icon themes.

However, it declares itself just as "Default", with no description.
Users can find this name confusing, and mistake it with the system's
default theme, so let's give it a better Name and Comment in our UI.

As a bonus we'll gain translations from our infrastructure which
the current theme doesn't provide.

BUG: 472953
FIXED-IN: 6