kcms/soundtheme: Add setting to disable notification sounds
Add a new setting option to control whether the notification
sounds are globally enabled. This affects application and
services that use KF6 KNotification.

It uses a checkbox instead of a switch because the setting is
not instant-apply.

When the option is disabled, so is the view to select themes,
as the theme will make no difference when no sound is played.
Besides any playing preview sound is stopped.