shell: set `QT_XCB_GL_INTEGRATION=xcb_egl` on X11
GLX is considered legacy, at this point. New functionality to expose zero-copy buffer passing is implemented as EGL extensions, for instance. The Mesa developers have recommended using EGL in place of GLX for a while, now.

The Wayland already have switched to EGL, so might as well bring the X11 forward.

Ref: https://gitlab.gnome.org/GNOME/gtk/-/merge_requests/3540