libtaskmanager: Improve handling of not yet mapped windows in region filter in TaskFilterProxyModel
If a new window is created but it's not mapped yet, kwin is going to
announce that window anyway.

On the other hand, the region filter in TaskFilterProxyModel checks
whether the window geometry is valid. If it's invalid, then it's going
to skip performing region checks. However, this may result in the filer
proxy model including windows that should not be there. For example,
this results in the panel losing its translucency temporarily when
launching apps because a filter proxy model is used to determine if any
window overlaps with the panel.

In order to fix that, TaskFilterProxyModel should ignore windows without
valid geometry yet.