applets/notifications: set Plasmoid.icon
The actual icon is drawn by CompactRepresentation, and currently it has
to be this way because it does a lot of complicated things including
animations, which can't be handled by Plasmoid.icon. However we should
set Plasmoid.icon anyway so that it can be used in other contexts that
read this property.

CCBUG: 473278