applets/systemtray: show the symbolic tray icons in the config window
This way what's shown in the config window will always match what's in
the tray, making it easy to visually associate them. It also means that
as we clean up our tray icons to always use symbolic icons (when
available), the config window's icons will follow suit and eventually
become 100% symbolic as well.

BUG: 473278
FIXED-IN: 6.0