Implement sorting of favorite runners
Counterpart of https://invent.kde.org/plasma/plasma-desktop/-/merge_requests/1610

To avoid the KRunner::ResultsModel having to know about the
implementation details how we set this config, a property is assigned to
the model.

This also helps with keeping default values (like the krunner_services
favorite) in sync with the plasma-desktop bits.

CCBUG: 340283