widgetexplorer: Simplify category filter creation
Instead of listing all applets again and comparing with the categories
of the already listed applets in the model just use the already known
categories.
Instead of using multiple sets and look ups if a filter was already added
only deduplicate the list of translated categories once.
Lowering of the category name is deferred to the filer/actual
comparison, this should make it less confusing what happens when and
we need the actual names for translation.
Also show the available filters sorted alphabetically by their translated
names.