ItemContainer: don't compile ConfigOverlay on plasmashell startup
During profiling plasmashell startup, I found sometimes
ConfigOverlay.qml complilation time can take about 50~100ms, which is
significant enough to slow down the startup. As it's not often for the
user to enter edit mode to edit widgets, only set the url on startup,
and load the config overlay on demand.