runners/services: Revalidate services being stale rather than reloading
For the most time during KRunners lifetime, one will not install new
apps or update existing ones. Thus is is far faster to only ensure if
the cache is valid rather than reloading all entries.

When reloading all entries, KService internally ensures that the ksycoca
cache is up to date.
Especially in KF6, we have less noise regarding the updated sycoca
cache, because services are no longer supported and were previously
used for config data.

Depends on https://invent.kde.org/frameworks/krunner/-/merge_requests/152
for a bugfix about the connections in AbstractRunner::init and
https://invent.kde.org/frameworks/kservice/-/merge_requests/159 for
making sure that we receive the signal in time.