Don't try to install waylandtasksmodeltest into the build dir
It already exists in the build dir and since CMAKE_BINARY_DIR holds the
full path it ends up trying to install into some ephemeral dir when
packaging for a distro, e.g.
"/var/tmp/paludis/build/kde-plasma-workspace-scm/work/build".

Instead just pass the proper path of the waylandtaskmodeltest to
waylandtasksmodeltestwrapper.py, which admittedly assumes searching for
a ECM version >= 5.38, but we already depend on a greater version.