logout-greeter: Add "Install Updates & Restart" button
When an offline update has been scheduled, turn the "Restart" button
into an "Install Updates & Restart" button and let the regular "Restart"
button cancel the pending update.

Also let the countdown timer hint to the fact that updates will be
installed on restart.

There's an explicit cancelSoftwareUpdateRequested() signal since a theme
might instead want to provide a simple hint with a "not now" button
that just cancels the pending update without restarting.

BUG: 435845