lockscreen: handle user switching also on nopassword page
this is a bit awkward but probably the best we can do without major
refactoring. previously we'd miss out on the user switch because that
only got handled on the regular login page, when dealing with a password
less account we'd get carried away by the succeeded signal to the
nopassword page and no user switching page would be on top of the stack
view.

the page flow has been slightly changed to better fit the use case:
- a new function maybeSwitchUser replaces the previous user switcher
- nopasswordunlock now **replaces** the pagestack (which is actually
what it did anyway because it has no back button)
- nopasswordunlock triggers maybeSwitchUser after completion
- regular mainview triggers maybeSwitchUser after completion

that end result is that in both scenarios we get a working switcher page

BUG: 458951
(cherry picked from commit 7d39865d8ce80380c29806641638fd45e52cbf01)