devicenotifier: Don't use one property as a safeguard to another
Propagation order of property change events is intentionally
non-deterministic, so depending on another property to provide safety
checks for nullable object access is an error-prone race condition.

One possible way around this limitation would be using functions, since
evaluating them in binding context would also bind to everything
accessed as with normal expressions. But in this case even a function
is not really needed, as a simple null check just works.