ksmserver: cap the amount of app restorations
when we restore more than ¼ of RLIMIT_NOFILE (default 1024) of a single
application it is incredibly likely that this application is being
defective (e.g. GUIless apps participating in session management but
also in autostarting -- resulting in ever multiplying instances).

when in excess, refuse to start **all** restorations of the application
(NOTE: this is why we loop everything twice)

CCBUG: 475506