Migrate session action shortcuts out of ksmserver
As part of a mission to make ksmserver only about X11 session
management, this patch splits the session actions out of ksmserver.

Making ksmserver optional at runtime is a pre-requisite for making
xwayland only load on demand.

This patch moves shortcuts to a KDED module with the same component
name to keep compatibilty