applets/systray: Fix alignment of entries config headers
The alignment of headers is based on mimicking the delegate's
fixed width for all columns but the "Entries" one which fills
the space.

Make sure that the width of the header matches that of the
list view so the alignment can work.

Also elade the text on "Entries" to avoid overlapping and
account for some extra spacings so the alignment to columns
is pixel perfect.

BUG: 476547
FIXED-IN: 6