libtaskmanager: Significantly speed up autotest
An autotest updates .desktop files and relies on ksycoca having the
updated files.

Ksycoca has an internal guard to ignore file changes that are within
1500ms.

The current test has a broken delay that always expires after 2.5
seconds, hiding the issue but always hitting the expiry timeout and
taking a long time to execute.

Instead we can use a hook for autotests to override ksycoca's internal
guard.