Panel settings redesign: plasmashell part
In the new design for the panel settings, the main window is now a PopupPlasmawindow
with the main controls, and an ancillary window which is a "panel" for the ruler view
visible only when the panel sizing is custom.

The contents and the UI for those will come from the shell package in plasma-desktop.

Supports https://invent.kde.org/plasma/plasma-desktop/-/merge_requests/1723.