libtaskmanager: Simplify taskmanager test
The taskmanager desktop entry test was written in a way that ran one iteration of the test in the main function, but to test variations ran iterations 2-5 inside a loop. 

The result was re-use of variables across different tests. This hit a bug because QProcess::terminate is non-blocking resulting in the test calling start on a running QProcess, which doesn't have the desired intention. 

By using a common code path for all iterations we can move some variables onto the stack simplifying the code and avoiding this issue. It also sets us up for a QTestData refactor.