appiumtests: Disable battery test due to circular dependency on powerdevil
The battery applet test mocks udev, but uses the real powerdevil.

powerdevil uses libkworkspace from plasma-workspace

This is a blocker for distros running these tests, and causes issues in
KDE's CI if we ever get out of sync.

We either need to mock powerdevil (instead of mocking udev) or have
intergration tests out of tree.