Fix global menu alignment and wrapping
The global menu was aligned to top by default, looking off with other elements.

I also added eliding so on panels that are too thin, the
text will elide and not "bleed" outside the panel.
Of course ideally the panel is wide enough to show the text nicely, but at least
it never escapes the panel.

Eliding also gives clear idea to the user that the panel is too thin to support this.

| Before | After |
| ------ | ------ |
| ![image](/uploads/a05ca5956c53836657b6db4ffee1cae0/image.png)       |     ![image](/uploads/13d13e74bb93003cef6b6a9bb3fe0641/image.png)   |