shutdown: Introduce support for KWin closing wayland windows
After ksmserver is done, KWin will now close (or not) open wayland
windows. Since this operation can also cancel the logout, ksmserver
gains a new hook to reset its state. This also means that it cannot
quit itself anymore but needs to be managed by plasma-shutdown.
Using this opportunity the  path for shutdown can be aligned between
X11 and Wayland. In the systemd-managed case we stop graphical-session.target,
otherwise plasma-shutdown quits KWin and ksmserver.
BUG: 461176
FIXED-IN: 6.0