Reopen running apps on startup as fallback session restore
Session restore is in a poor state. Not all applications have support,
especially with additional permissions needed for X11 on Flatpak.
On wayland we have nothing that works. Gnome's implementation is
non-standard, there's nothing in Qt6. There is definitely nothing that
can be done in time for 6.0.

Even if there was, we would have apps not supporting it.

This is a simplistic solution that relaunch any applications.

It's not a smart solution, there's nothing that brings back the exact
same session or handles multiple windows.
X11 apps can restore as before through ksmserver as before.

I'm still somewhat on the fence, but feedback so far has been positive.

This should not be merged without corresponding UI changes to not over
promise this as session restore.