reliably lookup preferred browser from scheme mimetype first
we have a number of options here with varying degrees of legacy.

- x-scheme-handler/http mimetype is the modern way to define a browser
- text/html is a previous dated way to define a browser
- kdeglobals BrowserApplication is an incredibly dated way

whenever we looked for one of them, now look for all three and do so in
the correct order. the code was also slightly reshuffled to return as
soon as possible when a match is found to ease reading