libkworkspace: Drop KDisplayManager introduce new simpler model
KDisplayManager is the oldest code in Plasma. It has abstractions for
things that haven't been active in over 20 years.

KDisplayManager's sole remaining job was handling of listing sessions
and switching between them. This list of sessions was then wrapped in a
SessionsModel for usage in screenlocker and the logout screen.

This patch keeps only the remaining active code for Logind and
ConsoleKit and exposes them directly as a model.

The code is synchronous to match the behaviour of existing clients, but
by being a model we can change that in the future without affecting ABI.