Use libxcrypt for salt generation
Many distros have already completed their migration to yescrypt for password hashes. For these distros it is desirable that the user KCM follows suit and generates yescrypt hashes for accountsservice by default.

The easiest way to handle this is to check to see if we're building with at least libxcrypt 4. This version is guaranteed to have the `crypt_gensalt` symbol which will typically default to yescrypt on those distros. The `crypt` function will detect that we're passing in a yescrypt prefix and generate the correct hash. This is functionally identical to how DEs that use libaccountsservice work.