applets/digital-clock: Fix padding for digital clock in oneLineDate state
The width taken by `timeLabel` is set to `sizehelper.contentWidth` which
does not reflect the real width of `timeLabel`. Setting it instead to
`timeLabel.paintedWidth` provides the actual rendered width of `timeLabel`.

`Grid` QML type already has the `columnSpacing` property set so we don't
need to manually add any spacing to the individual items.

BUG:476977