cmake: detatch GLIBC_LOCALE from environmental factors
it's now an explicit option that is on by default and must be explicitly
toggled off when it's not desired.

this ensures that all systems by default are getting a working locale
setup and can't accidentally lose it if we change conditions around. a
system without glibc meanwhile needs to opt out of this if it doesn't
support glibc-like locale definitions etc.