kcms/nightcolor: Add a graphic view of the day/night timings
Show the scheduled times and transitions for the day and night
lighting in a graphical view that portraits a 0-24h horizontal
axis, with icons for each state and a gradient with the actual
temperature colors for the transitions.

This view is placed on top to avoid visual jittering (going up
and down) between the different modes, and also clashing with
other visual elements, like the manual location map.

FEATURE: 414809
FIXED-IN: 6