libkworkspace: Use display manager directly for session switching
The current design is that workspace components requesting to switch
user will fire into the screensaver which then shows a UI of existing
sessions with an option to go to the display manager for a new session.

This is rather dated. Modern login managers re-use sessions when a user
logs in again. Logging in as the same user twice is fundamentally broken
due to the shared dbus-daemon anyway.

It's a simpler UX and code path to just always go back to the login
manager when we want to switch or create a new session.

The lockscreen maintains the switch user button. This button relocks
(which no-ops) and goes to the display manager. We don't need the
session listing page or to handle initally set flags.