kcms/nightcolor: Limit saved precision of geolocated coordinates
Two decimal places in the saved lat/lon coordinates are enough
to calculate the sunrise and sundown times, just as we do with
the manual location mode.

Otherwise, everytime the KCM is open on the "Automatic geolocation"
mode, there will be small differences between the saved and the
newly geolocated coords, making the KCM activate the "Reset" button.