kfontinst: don't crash on exit
lifetimes have changed in kf6: `widget()` is a child of `parent` and as
such a sibling to `CKCmFontInst` what happens at destruction is that
~parent runs -> ~widget runs (destroys its children - among them the
splitters) then ~CKCmFontInst runs -> crash.

Instead simply hold on to the destroyed signal of widget which fires
before its children (the splitters) get destroyed.

It feels a bit hacky but in the grands scheme of things the widget code
is not long for this world anyway.

Fixes SYSTEMSETTINGS-3A
BUG: 477001